use statmap init directly
fixes
