+++ bin/hdfcoinc/pycbc_coinc_statmap_inj
@@ -4,7 +4,7 @@ The program combines coincident output files generated
 by pycbc_coinc_findtrigs to generated a mapping between SNR and FAP, along
 with producing the combined foreground and background triggers
 """
-import argparse, h5py, logging, itertools, copy
+import argparse, h5py, logging, itertools, copy, pycbc.io
 from pycbc.future import numpy
 from itertools import izip
 from pycbc.events import veto, coinc
@@ -43,14 +43,14 @@ if args.verbose:
 
 logging.info("Loading coinc zerolag triggers")    
 window = args.cluster_window
-zdata = coinc.load_coincs(args.zero_lag_coincs)   
+zdata = pycbc.io.StatmapData(files=args.zero_lag_coincs)   
 zdata = zdata.cluster(window)
                      
 logging.info("Loading coinc full inj triggers")    
-fidata = coinc.load_coincs(args.mixed_coincs_full_inj).cluster(window)
+fidata = pycbc.io.StatmapData(files=args.mixed_coincs_full_inj).cluster(window)
                      
 logging.info("Loading coinc inj full triggers")    
-ifdata = coinc.load_coincs(args.mixed_coincs_inj_full).cluster(window)
+ifdata = pycbc.io.StatmapData(files=args.mixed_coincs_inj_full).cluster(window)
 
 f = h5py.File(args.output_file, "w")
